skyfield_sats/requirements.txt
numpy
skyfield
psycopg2-binary
hilbertcurve
